reorganizing files for user configuration
