+++ morAce/user/morse_code_shortcuts.py
@@ -1,0 +1,72 @@
+from adafruit_hid.keycode import Keycode
+
+morseCodeShortcutCmd = [
+  (".-",    "My favourite string here"),
+  ("-...",   "My password here"),
+  ("---.",  [Keycode.LEFT_CONTROL, Keycode.C]),
+  ("-..",   "My Password string here"),
+  (".",     [Keycode.E]),
+  ("..-.",  [Keycode.F]),
+  ("--.",   [Keycode.G]),
+  ("....",  [Keycode.LEFT_GUI, Keycode.H]),
+  ("..",    [Keycode.I]),
+  (".---",  [Keycode.J]),
+  ("-.-",   [Keycode.K]),
+  (".-..",  [Keycode.L]),
+  ("--",    [Keycode.M]),
+  ("-.",    [Keycode.LEFT_GUI, Keycode.TAB]),
+  ("---",   [Keycode.O]),
+  (".--.",  [Keycode.LEFT_CONTROL, Keycode.V]),
+  ("--.-",  [Keycode.Q]),
+  (".-.",   [Keycode.R]),
+  ("...",   [Keycode.S]),
+  ("-",     [Keycode.T]),
+  ("..-",   [Keycode.U]),
+  ("...-",  [Keycode.V]),
+  (".--",   [Keycode.W]),
+  ("-..-",  [Keycode.X]),
+  ("-.--",  [Keycode.Y]),
+  ("--..",  [Keycode.Z]),
+  ("-----", [Keycode.ZERO]),
+  (".----", [Keycode.ONE]),
+  ("..---", [Keycode.TWO]),
+  ("...--", [Keycode.THREE]),
+  ("....-", [Keycode.FOUR]),
+  (".....", [Keycode.FIVE]),
+  ("-....", [Keycode.SIX]),
+  ("--...", [Keycode.SEVEN]),
+  ("---..", [Keycode.EIGHT]),
+  ("----.", [Keycode.NINE]),
+  ("----..", [Keycode.BACKSLASH]),
+  ("....--", [Keycode.FORWARD_SLASH]),
+  (".--...", [Keycode.LEFT_BRACKET]),
+  ("-..---", [Keycode.RIGHT_BRACKET]),
+  ("--..--", [Keycode.SHIFT, Keycode.COMMA]),
+  ("..--..", [Keycode.SHIFT, Keycode.PERIOD]),
+  ("---...", [Keycode.SHIFT, Keycode.NINE]),
+  ("...---", [Keycode.SHIFT, Keycode.ZERO]),
+  ("--..-",  [Keycode.SHIFT, Keycode.LEFT_BRACKET]),
+  ("--..-",  [Keycode.SHIFT, Keycode.RIGHT_BRACKET]),
+  (".-----", [Keycode.PERIOD]),
+  ("-.....", [Keycode.COMMA]),
+  ("----.-", [Keycode.SHIFT, Keycode.MINUS]),
+  ("....-.", [Keycode.SHIFT, Keycode.BACKSLASH]),
+  ("-.----", [Keycode.SHIFT, Keycode.FORWARD_SLASH]),
+  (".-....", [Keycode.SHIFT, Keycode.ONE]),
+  ("-....-", [Keycode.SEMICOLON]),
+  (".----.", [Keycode.SHIFT, Keycode.SEMICOLON]),
+  (".---.",  [Keycode.MINUS]),
+  ("..----", [Keycode.SHIFT, Keycode.FOUR]),
+  ("...-.-", [Keycode.SHIFT, Keycode.FIVE]),
+  ("...--.", [Keycode.SHIFT, Keycode.QUOTE]),
+  ("---..-", [Keycode.SHIFT, Keycode.TWO]),
+  ("..-...", [Keycode.QUOTE]),
+  ("--.---", [Keycode.GRAVE_ACCENT]),
+  ("-...--", [Keycode.SHIFT, Keycode.SIX]),
+  ("---.--", [Keycode.SHIFT, Keycode.GRAVE_ACCENT]),
+  ("..---.", [Keycode.SHIFT, Keycode.THREE]),
+  (".---..", [Keycode.SHIFT, Keycode.SEVEN]),
+  ("-...-",  [Keycode.SHIFT,Keycode.EQUALS]),
+  ("---.-",  [Keycode.EQUALS]),
+  ("-..--",  [Keycode.SHIFT, Keycode.EIGHT])
+]
